🚚 [move] component > Landing 으로 이동
